requirements.txt
streamlit==0.58.0
pandas==1.0.1
joblib==0.14.1
matplotlib==3.1.3
numpy==1.18.1
Pillow==7.1.2
